Unify sync process for root activities (lists) and sync tasks #213
pushEntities' calls to rest interface should validate the result. The server should return the number of received dtos.
handle synchronization errors -> RetroFacade needs to disconnect and reinitialize

+++ sormas-app/app/src/main/java/de/symeda/sormas/app/backend/caze/CaseDtoHelper.java
@@ -60,7 +60,7 @@ public class CaseDtoHelper extends AdoDtoHelper<Case, CaseDataDto> {
     }
 
     @Override
-    protected Call<Long> pushAll(List<CaseDataDto> caseDataDtos) {
+    protected Call<Integer> pushAll(List<CaseDataDto> caseDataDtos) {
         return RetroProvider.getCaseFacade().pushAll(caseDataDtos);
     }
