+++ sormas-app/app/src/main/java/de/symeda/sormas/app/backend/symptoms/SymptomsDtoHelper.java
@@ -33,7 +33,7 @@ public class SymptomsDtoHelper extends AdoDtoHelper<Symptoms, SymptomsDto> {
     }
 
     @Override
-    protected Call<Long> pushAll(List<SymptomsDto> symptomsDtos) {
+    protected Call<Integer> pushAll(List<SymptomsDto> symptomsDtos) {
         throw new UnsupportedOperationException("Entity is embedded");
     }
