+++ sormas-app/app/src/main/java/de/symeda/sormas/app/rest/TaskFacadeRetro.java
@@ -20,7 +20,7 @@ public interface TaskFacadeRetro {
     Call<List<TaskDto>> pullAllSince(@Path("since") long since);
 
     @POST("tasks/push")
-    Call<Long> pushAll(@Body List<TaskDto> dtos);
+    Call<Integer> pushAll(@Body List<TaskDto> dtos);
 
     @GET("tasks/uuids")
     Call<List<String>> pullUuids();
